remove unecesary listener and add the code on auto route listener
